Add files via upload
This is a review of Random Financial stocks I selected from Yahoo! Finance because I wanted to try my hand at Time Analysis Visualization.

Stock prices constantly change as a function of time and volume and those variables affects the direction it's prices travel in. I used Python language and the Pandas library to not only convert Date-Time values to something Python could understand, but also an easier view of the path the stock has traveled within the past year.